crawler-manager/requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
asyncio==3.4.3
atomicwrites==1.3.0
attrs==18.2.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cssselect==1.0.3
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.3.2
MarkupSafe==1.1.1
more-itertools==6.0.0
multidict==4.5.2
mysql-connector-python==8.0.15
parsel==1.5.1
pluggy==0.9.0
protobuf==3.7.0
py==1.8.0
pytest==4.3.0
redis==3.2.0
requests==2.21.0
six==1.12.0
urllib3==1.24.1
w3lib==1.20.0
Werkzeug==0.14.1
yarl==1.3.0